Move application logic from main.go to app.go
